refactor(dsolve): move Factorable tests to test_single.py
